fix: css prevented desktop mode to render correctly
The issue was that the scrollbar was taken into account in the vw
variable, to the sidebar and main content were too large to fit next to
each other.

The solution is to position the parent div in a fixed position. This
prevent the body from having a scrollbar, so the vw width is always the
same.